doremi_subscriptions/requirements.txt
PyYAML>=6.0
python-dateutil>=2.8.2